requirements.txt
Pillow==8.3.2
PyQt5==5.15.4
configparser==5.0.2